Adjust assertion verification code
Currently, attempting to use Google's IdP results in SAMLY stalling after
consuming the SAML assertions and replying with "access_denied :bad_assertion".

This is because the assertion checking code begins searching at the wrong
location in the XML, so tweak that to work. This might not be the best
solution.

See https://github.com/arekinath/esaml/issues/17